+++ packages/sdk/src/commonTypes.ts
@@ -1,9 +1,9 @@
 import { Action } from "./contract";
 import { JoinSplitInputs } from "./proof/joinsplit";
-import { CanonAddress, StealthAddress } from "./crypto/address";
+import { StealthAddress } from "./crypto/address";
 import { BaseProof, MerkleProofInput } from "./proof";
 import { IncludedNote, Note } from "./sdk/note";
-import { Asset, EncodedAsset } from "./sdk/asset";
+import { EncodedAsset } from "./sdk/asset";
 
 export const SNARK_SCALAR_FIELD =
   21888242871839275222246405745257275088548364400416034343698204186575808495617n;
@@ -12,32 +12,6 @@ export const BLOCK_GAS_LIMIT = 30_000_000n;
 export type Address = string;
 export type NoteAssetKey = string; // Takes form of NOTES_<address>_<id>
 export type AllNotes = Map<NoteAssetKey, IncludedNote[]>;
-
-export interface UnwrapRequest {
-  asset: Asset;
-  unwrapValue: bigint;
-}
-
-export interface PaymentIntent {
-  receiver: CanonAddress;
-  value: bigint;
-}
-
-// A joinsplit request is an unwrapRequest plus an optional payment
-export interface JoinSplitRequest extends UnwrapRequest {
-  paymentIntent?: PaymentIntent;
-}
-
-export interface OperationRequest {
-  joinSplitRequests: JoinSplitRequest[];
-  refundAddr?: StealthAddress;
-  refundAssets: Asset[];
-  actions: Action[];
-  verificationGasLimit?: bigint;
-  executionGasLimit?: bigint;
-  maxNumRefunds?: bigint;
-  gasPrice?: bigint;
-}
 
 export type SolidityProof = [
   bigint,